5464: fix build errors
